Libs/uwb-stack/config_manager/tools/config_plugin_gen/pyproject.toml
[tool.poetry]
name = "config-plugin-gen"
version = "0.0.1"
description = "Tool to manage the Config Manager."
repository = "https://gitlab.com/qorvo"
authors = ["Qorvo Inc <qorvo.inc@qorvo.com>"]
include = ["config_plugin_gen/templates/*"]

[tool.poetry.dependencies]
python = "^3.9.0"
jinja2 = "^3.1.2"
click = "^8.1.3"
loguru = "^0.6.0"
PyYAML = "^6.0"
pydantic-yaml = "^1.2.0"
pydantic = "^2.5.2"
ttable = "^0.6.4"
expr-py = "^0.3.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.1"
flake8-pyproject = "^1.2.2"
flake8 = "^6.0.0"
black = "^23.1.0"
flake8-docstrings = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
config-plugin-gen = "config_plugin_gen.cli:cli_root"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
ignore_missing_imports = true
follow_imports = "skip"
strict_optional = true
plugins = ["pydantic.mypy"]
warn_redundant_casts = "True"
warn_unused_ignores = "True"
disallow_any_generics = "True"
check_untyped_defs = "True"
no_implicit_reexport = "True"
disallow_untyped_defs = "True"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.flake8]
docstring-convention = "numpy"
exclude = [
	".cmake-format.py"
]
max-line-length = 100
